refactor: remove schemas folder
