+++ src/charts/dual-axes.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { zodToJsonSchema } from "zod-to-json-schema";
-import { BaseConfigSchema } from "../base";
+import { BaseConfigSchema } from "./base";
 
 // Dual axes series schema
 const DualAxesSeriesSchema = z.object({
