+++ src/charts/treemap.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { zodToJsonSchema } from "zod-to-json-schema";
-import { BaseConfigSchema } from "../base";
+import { BaseConfigSchema } from "./base";
 
 // Define recursive schema for hierarchical data
 const TreeNodeSchema: z.ZodType<any> = z.lazy(() =>
